requirements.txt
ujson
pyyaml
uvloop
aiogram
apscheduler